Create ex1data2-3features.txt
This file implements gender a different way, for the 5K race results problem.  

Gender is split into two features ("man?" and "woman?", each of which is either 0 or 1).   Therefore the multivariate linear regression is to find a theta for each of these features, plus the person's age.  

"woman?" is column 2, and "man?" is column 3.